requirements.txt
transformers-stream-generator
qwen-vl-utils
docling-core
transformers
torchvision
accelerate
matplotlib
easydict
requests
hf_xet
spaces
pillow
gradio
addict
einops
torch
numpy
timm